[SYCL] Add KHR builtin include wrappers with regression coverage
Refactors builtins.hpp into a thin hub plus builtin_helpers.hpp
and per-category internal headers. Intended to improve header
organization and introduce opportunities for reducing compile-time cost.

Introduces KHR inclusion points to the fine grain includes
