+++ sycl/include/sycl/builtins_common.hpp
@@ -1,0 +1,13 @@
+//==-------- builtins_common.hpp - SYCL built-in common functions --------==//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#pragma once
+
+#include <sycl/detail/defines_elementary.hpp>
+
+#include <sycl/detail/builtins/common_functions.hpp>
