+++ sycl/include/sycl/builtins_math.hpp
@@ -1,0 +1,15 @@
+//==-------- builtins_math.hpp - SYCL built-in math functions ------------==//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#pragma once
+
+#include <sycl/detail/defines_elementary.hpp>
+
+#include <sycl/detail/builtins/half_precision_math_functions.hpp>
+#include <sycl/detail/builtins/math_functions.hpp>
+#include <sycl/detail/builtins/native_math_functions.hpp>
