+++ sycl/include/sycl/detail/builtins/integer_functions.hpp
@@ -6,8 +6,9 @@
 //
 //===----------------------------------------------------------------------===//
 
-// Intentionally insufficient set of includes and no "#pragma once".
+#pragma once
 
+#include <sycl/detail/builtins/builtin_helpers.hpp>
 #include <sycl/detail/builtins/helper_macros.hpp>
 
 namespace sycl {
@@ -214,8 +215,7 @@ DEVICE_IMPL_TEMPLATE(TWO_ARGS, mul24, enable_upsample_t, [](auto... xs) {
 #else
 template <typename T0, typename T1>
 std::enable_if_t<detail::enable_upsample_v<T0, T1>,
-                 detail::upsample_ret_type_t<T0>>
-upsample(T0 x, T1 y) {
+                 detail::upsample_ret_type_t<T0>> upsample(T0 x, T1 y) {
   using namespace detail;
   if constexpr (is_vec_or_swizzle_v<T0> || is_marray_v<T0>) {
     return builtin_delegate_to_scalar(
