+++ sycl/include/sycl/khr/includes/builtins_relational.hpp
@@ -1,0 +1,16 @@
+//===----------------------------------------------------------------------===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef __SYCL_KHR_INCLUDES_BUILTINS_RELATIONAL
+#define __SYCL_KHR_INCLUDES_BUILTINS_RELATIONAL
+
+#include "version.hpp"
+
+#include <sycl/builtins_relational.hpp>
+
+#endif // __SYCL_KHR_INCLUDES_BUILTINS_RELATIONAL
