+++ sycl/test/include_deps/sycl_khr_includes_stream.hpp.cpp
@@ -98,19 +98,20 @@
 // CHECK-NEXT: __spirv/spirv_ops.hpp
 // CHECK-NEXT: builtins.hpp
 // CHECK-NEXT: detail/builtins/builtins.hpp
+// CHECK-NEXT: detail/builtins/common_functions.hpp
+// CHECK-NEXT: detail/builtins/builtin_helpers.hpp
 // CHECK-NEXT: detail/vector_core.hpp
 // CHECK-NEXT: detail/named_swizzles_mixin.hpp
 // CHECK-NEXT: detail/vector_traits.hpp
 // CHECK-NEXT: detail/memcpy.hpp
 // CHECK-NEXT: marray.hpp
-// CHECK-NEXT: detail/builtins/common_functions.inc
 // CHECK-NEXT: detail/builtins/helper_macros.hpp
-// CHECK-NEXT: detail/builtins/geometric_functions.inc
-// CHECK-NEXT: detail/builtins/half_precision_math_functions.inc
-// CHECK-NEXT: detail/builtins/integer_functions.inc
-// CHECK-NEXT: detail/builtins/math_functions.inc
-// CHECK-NEXT: detail/builtins/native_math_functions.inc
-// CHECK-NEXT: detail/builtins/relational_functions.inc
+// CHECK-NEXT: detail/builtins/geometric_functions.hpp
+// CHECK-NEXT: detail/builtins/half_precision_math_functions.hpp
+// CHECK-NEXT: detail/builtins/integer_functions.hpp
+// CHECK-NEXT: detail/builtins/math_functions.hpp
+// CHECK-NEXT: detail/builtins/native_math_functions.hpp
+// CHECK-NEXT: detail/builtins/relational_functions.hpp
 // CHECK-NEXT: group.hpp
 // CHECK-NEXT: detail/async_work_group_copy_ptr.hpp
 // CHECK-NEXT: device_event.hpp
